+++ packages/framework/lib/src/core/response_context.dart
@@ -287,7 +287,7 @@ abstract class ResponseContext<RawResponse>
     }
 
     final head = controller
-        .findExpose(app!.container!.reflector)!
+        .findExpose(app!.container.reflector)!
         .path
         .toString()
         .replaceAll(_straySlashes, '');