+++ packages/production/example/main.dart
@@ -8,15 +8,15 @@ void main(List<String> args) => Runner('example', configureServer).run(args);
 
 Future configureServer(Angel app) async {
   // Use the injected `pub_sub.Client` to send messages.
-  var client = app.container?.make<pub_sub.Client>();
+  var client = app.container.make<pub_sub.Client>();
   var greeting = 'Hello! This is the default greeting.';
 
   // We can listen for an event to perform some behavior.
   //
   // Here, we use message passing to synchronize some common state.
-  var onGreetingChanged = await client?.subscribe('greeting_changed');
+  var onGreetingChanged = await client.subscribe('greeting_changed');
   onGreetingChanged
-      ?.cast<String>()
+      .cast<String>()
       .listen((newGreeting) => greeting = newGreeting);
 
   // Add some routes...
@@ -33,7 +33,7 @@ Future configureServer(Angel app) async {
   // This route will push a new value for `greeting`.
   app.get('/change_greeting/:newGreeting', (req, res) {
     greeting = (req.params['newGreeting'] as String? ?? '');
-    client?.publish('greeting_changed', greeting);
+    client.publish('greeting_changed', greeting);
     return 'Changed greeting -> $greeting';
   });